requirements.txt

# ── Core ML / Embeddings ──────────────────────────────────────
sentence-transformers>=2.7.0,<3.0.0
transformers>=4.40.0,<5.0.0
torch>=2.2.0

# ── Vector Search ─────────────────────────────────────────────
faiss-cpu>=1.8.0

# ── LangChain stack ───────────────────────────────────────────
langchain>=0.2.0,<0.4.0
langchain-core>=0.2.0,<0.4.0
langchain-community>=0.2.0,<0.4.0
langchain-ollama>=0.1.0

# ── DuckDuckGo search
# FIX: the package is still called duckduckgo-search (not ddgs which doesn't exist on PyPI)
duckduckgo-search>=6.0.0

# ── NLP / PII Detection ───────────────────────────────────────
spacy>=3.7.0,<4.0.0
# REQUIRED after install: python -m spacy download en_core_web_sm
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0

# ── Toxicity Classifier ───────────────────────────────────────
detoxify>=0.5.0

# ── Dataset Loading ───────────────────────────────────────────
datasets>=2.19.0

# ── API Layer ─────────────────────────────────────────────────
flask>=3.0.0
flask-cors>=4.0.0

# ── Utilities ─────────────────────────────────────────────────
pyyaml>=6.0
python-dotenv>=1.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
pandas>=2.2.0
tqdm>=4.66.0
